requirements.txt
PyYAML==6.0.1
openai==1.3.0
elevenlabs==0.2.26
requests==2.31.0
